+++ authservice/auth_service.go
@@ -15,7 +15,6 @@
 package authservice
 
 import (
-	"bytes"
 	"crypto/rsa"
 	"encoding/json"
 	"errors"
@@ -299,32 +298,15 @@ type RemoteServiceAccountLoaderConfig struct {
 	AccessTokenPath     string // Path to auth service access token endpoint
 	DeletedAccountsPath string // Path to auth service deleted accounts endpoint
 
+	AccessTokenRequest       *http.Request
 	DeletedAccountsCallback  func([]string) error // Function to call once the deleted accounts list is received from the auth service
 	GetDeletedAccountsPeriod int64                // How often to request deleted account list from the auth service (in hours)
 }
 
-// GetAccessToken implements ServiceAccountLoader interface
+// GetAccessToken implements AuthDataLoader interface
 func (r *RemoteServiceAccountLoaderImpl) GetAccessToken() error {
-	params := map[string]interface{}{
-		"auth_type": "static_token",
-		"creds": map[string]string{
-			"token": r.config.ServiceToken,
-		},
-	}
-	data, err := json.Marshal(params)
-	if err != nil {
-		return fmt.Errorf("error marshaling request body toget access token: %v", err)
-	}
-
 	client := &http.Client{}
-	req, err := http.NewRequest("POST", r.config.AuthServicesHost+r.config.AccessTokenPath, bytes.NewReader(data))
-	if err != nil {
-		return fmt.Errorf("error formatting request to get access token: %v", err)
-	}
-
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := client.Do(req)
+	resp, err := client.Do(r.config.AccessTokenRequest)
 	if err != nil {
 		return fmt.Errorf("error requesting access token: %v", err)
 	}
@@ -450,7 +432,11 @@ func NewRemoteServiceAccountLoader(config RemoteServiceAccountLoaderConfig, logg
 	if config.ServiceToken == "" && config.DeletedAccountsCallback != nil {
 		return nil, errors.New("service token is missing")
 	}
-	constructServiceAccountLoaderConfig(&config)
+
+	err := constructServiceAccountLoaderConfig(&config)
+	if err != nil {
+		return nil, fmt.Errorf("error constructing data loader config: %v", err)
+	}
 
 	timerDone := make(chan bool)
 
@@ -463,16 +449,26 @@ func NewRemoteServiceAccountLoader(config RemoteServiceAccountLoaderConfig, logg
 	return &dataLoader, nil
 }
 
-func constructServiceAccountLoaderConfig(config *RemoteServiceAccountLoaderConfig) {
+func constructServiceAccountLoaderConfig(config *RemoteServiceAccountLoaderConfig) error {
 	if config.AccessTokenPath == "" {
 		config.AccessTokenPath = "/bbs/access-token"
 	}
 	if config.DeletedAccountsPath == "" {
 		config.DeletedAccountsPath = "/bbs/deleted-accounts"
 	}
+	if config.AccessTokenRequest == nil {
+		r, err := authutils.GetDefaultAccessTokenRequest(config.AuthServicesHost, config.AccessTokenPath, config.ServiceToken)
+		if err != nil {
+			return err
+		}
+
+		config.AccessTokenRequest = r
+	}
 	if config.GetDeletedAccountsPeriod <= 0 {
 		config.GetDeletedAccountsPeriod = 2
 	}
+
+	return nil
 }
 
 // AccessToken represents an access token granted by a remote auth service