+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - Split AuthDataLoader interface [#36](https://github.com/rokwire/core-auth-library-go/issues/36)
+### Added
+- Add support for signature auth to retrieve access tokens [#26](https://github.com/rokwire/core-auth-library-go/issues/26)
 
 ## [1.0.6] - 2022-03-03
 ### Changed